% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{new.logger}
\alias{new.logger}
\title{new logger}
\usage{
new.logger(logfile)
}
\arguments{
\item{logfile}{the name of the log file}
}
\value{
a new logger object with the following methods:
\describe{
  \item{info(...)}{writes an info message to the log file}
  \item{warn(...)}{writes an warning message to the log file}
  \item{err(...)}{writes an error message to the log file}
}
}
\description{
Constructs a new logger object. Allows for writing to log files and stdout. 
Entries are tagged by time stamps and event type.
}
\examples{
logger <- new.logger("mylogfile.log")
logger$info("This is a test log entry!")

}
