% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{registerLogErrorHandler}
\alias{registerLogErrorHandler}
\title{register log-enabled error handler}
\usage{
registerLogErrorHandler(logger, die = TRUE)
}
\arguments{
\item{logger}{a yogilogger object to be used for logging any fatal errors}

\item{die}{a logical value indicationg whether the process should exit with an error status
upon handling the error. Defaults to TRUE}
}
\description{
Registers a top-level error handler that makes a log entry and exiting with an error status
}
\details{
When an error occurs, the handler will catch it, make a log entry detailing the error message
and stack trace, and then exit with an error status (unless overridden with die=FALSE)
}
\examples{

}
